Refactor useCopyPagePages into more files as it gets too long
